fix: unittest has been modified to test refactored code
